+++ src/main/java/com/imsweb/algorithms/napiia/NapiiaInputPatientDto.java
@@ -3,17 +3,105 @@
  */
 package com.imsweb.algorithms.napiia;
 
-import java.util.List;
-
 public class NapiiaInputPatientDto {
-    
-    List<NapiiaInputRecordDto> _napiiaInputPatientDtoList;
 
-    public List<NapiiaInputRecordDto> getNapiiaInputPatientDtoList() {
-        return _napiiaInputPatientDtoList;
+    private String _race1;
+
+    private String _race2;
+
+    private String _race3;
+
+    private String _race4;
+
+    private String _race5;
+
+    private String _spanishHispanicOrigin;
+
+    private String _birthplaceCountry;
+
+    private String _sex;
+
+    private String _nameLast;
+
+    private String _nameBirthSurname;
+
+    public String getRace1() {
+        return _race1;
     }
 
-    public void setNapiiaInputPatientDtoList(List<NapiiaInputRecordDto> napiiaInputPatientDtoList) {
-        _napiiaInputPatientDtoList = napiiaInputPatientDtoList;
+    public void setRace1(String race1) {
+        _race1 = race1;
+    }
+
+    public String getRace2() {
+        return _race2;
+    }
+
+    public void setRace2(String race2) {
+        _race2 = race2;
+    }
+
+    public String getRace3() {
+        return _race3;
+    }
+
+    public void setRace3(String race3) {
+        _race3 = race3;
+    }
+
+    public String getRace4() {
+        return _race4;
+    }
+
+    public void setRace4(String race4) {
+        _race4 = race4;
+    }
+
+    public String getRace5() {
+        return _race5;
+    }
+
+    public void setRace5(String race5) {
+        _race5 = race5;
+    }
+
+    public String getSpanishHispanicOrigin() {
+        return _spanishHispanicOrigin;
+    }
+
+    public void setSpanishHispanicOrigin(String spanishHispanicOrigin) {
+        _spanishHispanicOrigin = spanishHispanicOrigin;
+    }
+
+    public String getBirthplaceCountry() {
+        return _birthplaceCountry;
+    }
+
+    public void setBirthplaceCountry(String birthplaceCountry) {
+        _birthplaceCountry = birthplaceCountry;
+    }
+
+    public String getSex() {
+        return _sex;
+    }
+
+    public void setSex(String sex) {
+        _sex = sex;
+    }
+
+    public String getNameLast() {
+        return _nameLast;
+    }
+
+    public void setNameLast(String nameLast) {
+        _nameLast = nameLast;
+    }
+
+    public String getNameBirthSurname() {
+        return _nameBirthSurname;
+    }
+
+    public void setNameBirthSurname(String nameBirthSurname) {
+        _nameBirthSurname = nameBirthSurname;
     }
 }